\name{spatstat.utils-deprecated} 
\title{Deprecated Functions of spatstat.utils Package}
\alias{spatstat.utils-deprecated}  %DoNotExport
\alias{equispaced}          
\description{
  Deprecated functions of the \code{spatstat.utils} package.
}
\usage{
equispaced(z, reltol=0.001)
}
\arguments{
  \item{z}{Numeric vector.}
  \item{reltol}{Numeric value for relative tolerance.}
}
\details{
  \code{equispaced} has been replaced by \code{\link{evenly.spaced}}.
}
\keyword{deprecated}
